+++ src/main/kotlin/me/fzzyhmstrs/fzzy_config/config/ConfigGroup.kt
@@ -30,12 +30,8 @@ import net.minecraft.client.gui.DrawContext
 import net.minecraft.client.gui.screen.narration.NarrationMessageBuilder
 import net.minecraft.client.gui.screen.narration.NarrationPart
 import net.minecraft.client.gui.tooltip.Tooltip
-import net.minecraft.screen.ScreenTexts
-import net.minecraft.text.OrderedText
-import net.minecraft.text.StringVisitable
 import net.minecraft.text.Text
 import net.minecraft.util.Identifier
-import net.minecraft.util.Language
 import org.jetbrains.annotations.ApiStatus.Internal
 import java.util.*