+++ src/main/kotlin/me/fzzyhmstrs/fzzy_config/util/platform/impl/PlatformApiImpl.kt
@@ -13,10 +13,12 @@ package me.fzzyhmstrs.fzzy_config.util.platform.impl
 import me.fzzyhmstrs.fzzy_config.FC
 import me.fzzyhmstrs.fzzy_config.util.PlatformApi
 import me.fzzyhmstrs.fzzy_config.util.platform.Registrar
+import net.fabricmc.loader.api.FabricLoader
 import net.fabricmc.loader.api.Version
 import net.minecraft.registry.Registry
 import org.slf4j.Logger
 import java.io.File
+import java.util.*
 
 internal object PlatformApiImpl: PlatformApi {
 
@@ -48,11 +50,11 @@ internal object PlatformApiImpl: PlatformApi {
         return RegistrarImpl(namespace, registry)
     }
 
-    fun testVersion(id: String, version: String): Optional<Int> {
+    override fun testVersion(id: String, version: String): Optional<Int> {
         return try {
             FabricLoader.getInstance().getModContainer(id).map { container ->
                 val v = Version.parse(version)
-                container.getMetadata().getVersion().compareTo(v)
+                container.metadata.version.compareTo(v)
             }
         } catch (e: Throwable) {
             FC.DEVLOG.error("Critical error encountered parsing version in PlatformApi", e)