+++ src/main/kotlin/me/fzzyhmstrs/fzzy_config/validation/collection/ListListWidget.kt
@@ -13,8 +13,8 @@ package me.fzzyhmstrs.fzzy_config.validation.collection
 import me.fzzyhmstrs.fzzy_config.entry.EntryValidator
 import me.fzzyhmstrs.fzzy_config.screen.internal.SuggestionWindowListener
 import me.fzzyhmstrs.fzzy_config.screen.internal.SuggestionWindowProvider
-import me.fzzyhmstrs.fzzy_config.screen.widget.TextlessActionWidget
 import me.fzzyhmstrs.fzzy_config.screen.widget.TextureIds
+import me.fzzyhmstrs.fzzy_config.screen.widget.custom.CustomButtonWidget
 import me.fzzyhmstrs.fzzy_config.util.ValidationResult
 import me.fzzyhmstrs.fzzy_config.validation.misc.ChoiceValidator
 import net.minecraft.client.MinecraftClient
@@ -102,18 +102,19 @@ internal class ListListWidget<T>(entryList: List<me.fzzyhmstrs.fzzy_config.entry
         private var clickedWidget: Element? = null
 
         private val entryWidget = entry.widgetAndTooltipEntry(validator.apply(parent, this)).also { if (it is SuggestionWindowProvider) it.addListener(parent) }
-        private val deleteWidget = TextlessActionWidget(
-            TextureIds.DELETE,
-            TextureIds.DELETE_INACTIVE,
-            TextureIds.DELETE_HIGHLIGHTED,
-            TextureIds.DELETE_LANG,
-            TextureIds.DELETE_LANG,
-            { true },
-            { parent.children().let { list ->
+
+        private val deleteWidget = CustomButtonWidget.builder { parent.children().let { list ->
                 list.indexOf(this).takeIf { i -> i >=0 && i<list.size }?.let {
                         i -> list.removeAt(i)
                 }
-            } })
+            } }
+            .textures(TextureIds.DELETE,
+                TextureIds.DELETE_INACTIVE,
+                TextureIds.DELETE_HIGHLIGHTED)
+            .tooltip(TextureIds.DELETE_LANG)
+            .narrationSupplier { _, _ -> TextureIds.DELETE_LANG }
+            .size(20, 20)
+            .build()
 
         fun get(): T {
             return entry.get()
@@ -161,18 +162,17 @@ internal class ListListWidget<T>(entryList: List<me.fzzyhmstrs.fzzy_config.entry
     @Suppress("UNCHECKED_CAST")
     private class NewEntry<T>(private val entrySupplier: me.fzzyhmstrs.fzzy_config.entry.Entry<T, *>, private val parent: ListListWidget<T>, private val validator: BiFunction<ListListWidget<T>, ListEntry<T>?, ChoiceValidator<T>>): ListEntry<T>() {
 
-        private val addWidget = TextlessActionWidget(
-            TextureIds.ADD,
-            TextureIds.ADD_INACTIVE,
-            TextureIds.ADD_HIGHLIGHTED,
-            TextureIds.ADD_LANG,
-            TextureIds.ADD_LANG,
-            { true },
-            {
+        private val addWidget = CustomButtonWidget.builder {
                 parent.children().let { it.add(it.lastIndex, ExistingEntry(entrySupplier.instanceEntry() as me.fzzyhmstrs.fzzy_config.entry.Entry<T, *>, parent, validator)) }
                 parent.makeVisible(this)
-            })
-
+            }
+            .textures(TextureIds.ADD,
+                TextureIds.ADD_INACTIVE,
+                TextureIds.ADD_HIGHLIGHTED)
+            .tooltip(TextureIds.ADD_LANG)
+            .narrationSupplier { _, _ -> TextureIds.ADD_LANG }
+            .size(20, 20)
+            .build()
 
         override fun children(): MutableList<out Element> {
             return mutableListOf(addWidget)