+++ src/main/kotlin/me/fzzyhmstrs/fzzy_config/validation/collection/ValidatedIdentifierMap.kt
@@ -16,6 +16,7 @@ import me.fzzyhmstrs.fzzy_config.entry.EntryValidator
 import me.fzzyhmstrs.fzzy_config.impl.ConfigApiImpl
 import me.fzzyhmstrs.fzzy_config.screen.decoration.Decorated
 import me.fzzyhmstrs.fzzy_config.screen.widget.*
+import me.fzzyhmstrs.fzzy_config.screen.widget.custom.CustomButtonWidget
 import me.fzzyhmstrs.fzzy_config.util.ValidationResult
 import me.fzzyhmstrs.fzzy_config.util.ValidationResult.Companion.report
 import me.fzzyhmstrs.fzzy_config.validation.ValidatedField
@@ -185,7 +186,7 @@ open class ValidatedIdentifierMap<V>(defaultValue: Map<Identifier, V>, private v
     @Internal
     //client
     override fun widgetEntry(choicePredicate: ChoiceValidator<Map<Identifier, V>>): ClickableWidget {
-        return ActiveButtonWidget(TextureIds.MAP_LANG, 110, 20, { true }, { b: ActiveButtonWidget -> openMapEditPopup(b) })
+        return CustomButtonWidget.builder(TextureIds.MAP_LANG) { b: CustomButtonWidget -> openMapEditPopup(b) }.size(110, 20).build()
     }
 
     @Internal
@@ -195,7 +196,7 @@ open class ValidatedIdentifierMap<V>(defaultValue: Map<Identifier, V>, private v
 
     @Suppress("UNCHECKED_CAST")
     //client
-    private fun openMapEditPopup(b: ActiveButtonWidget) {
+    private fun openMapEditPopup(b: CustomButtonWidget) {
         try {
             val map = storedValue.map {
                 Pair(