+++ src/main/kotlin/me/fzzyhmstrs/fzzy_config/validation/collection/ValidatedMap.kt
@@ -20,6 +20,7 @@ import me.fzzyhmstrs.fzzy_config.screen.context.ContextResultBuilder
 import me.fzzyhmstrs.fzzy_config.screen.context.ContextType
 import me.fzzyhmstrs.fzzy_config.screen.decoration.Decorated
 import me.fzzyhmstrs.fzzy_config.screen.widget.*
+import me.fzzyhmstrs.fzzy_config.screen.widget.custom.CustomButtonWidget
 import me.fzzyhmstrs.fzzy_config.util.FcText.translate
 import me.fzzyhmstrs.fzzy_config.util.ValidationResult
 import me.fzzyhmstrs.fzzy_config.util.ValidationResult.Companion.report
@@ -205,7 +206,7 @@ open class ValidatedMap<K, V>(defaultValue: Map<K, V>, private val keyHandler: E
     @Internal
     //client
     override fun widgetEntry(choicePredicate: ChoiceValidator<Map<K, V>>): ClickableWidget {
-        return ActiveButtonWidget("fc.validated_field.map".translate(), 110, 20, { true }, { b: ActiveButtonWidget -> openMapEditPopup(b) })
+        return CustomButtonWidget.builder(TextureIds.MAP_LANG) { b: CustomButtonWidget -> openMapEditPopup(b) }.size(110, 20).build()
     }
 
     @Internal
@@ -226,7 +227,7 @@ open class ValidatedMap<K, V>(defaultValue: Map<K, V>, private val keyHandler: E
 
     @Suppress("UNCHECKED_CAST")
     //client
-    private fun openMapEditPopup(b: ActiveButtonWidget) {
+    private fun openMapEditPopup(b: CustomButtonWidget) {
         try {
             val map = storedValue.map {
                 Pair(